feat: Add new field for send the new loan monthly to lambda function and create enum for status loan

+++ domain/model/src/main/java/co/com/loanapplications/model/loanapplication/events/LoanSummary.java
@@ -5,12 +5,14 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class LoanSummary {
     private Double principal;
-    private Double annualInterestRate;
+    private BigDecimal annualInterestRate;
     private Integer termMonths;
 }
